ダイアログの実装

+++ src/app/component/page/pick-up/pick-up.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { StorageService } from 'src/app/service/storage/storage.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Yotei } from 'src/app/interface/yotei';
+import { MatDialog } from '@angular/material/dialog';
+import { AlertDialogComponent } from 'src/app/component/parts/alert-dialog/alert-dialog.component';
 
 @Component({
   selector: 'app-pick-up',
@@ -13,6 +15,7 @@ export class PickUpComponent implements OnInit {
   public yotei: Yotei[];
 
   constructor(
+    private dialog: MatDialog,
     private storageService: StorageService,
     private snackBar: MatSnackBar,
   ) {
@@ -27,17 +30,32 @@ export class PickUpComponent implements OnInit {
    * データを削除
    */
   public deleteAllData() {
-    this.storageService.deleteAllData();
+    const dialogRef = this.dialog.open(AlertDialogComponent, {
+      width: '400px',
+      height: '250px',
+      data: {
+        title: 'Pickデータの削除',
+        message: 'Pickしたデータを全て削除してもよろしいでしょうか？'
+      },
+    });
 
-    // pickした予定データを全て空にする
-    this.yotei = [];
+    dialogRef.afterClosed().subscribe(result => {
+      // ダイアログでキャンセル押下時、削除処理は実行しない
+      if (!result) return;
 
-    // スナックバー表示
-    this.snackBar.open("全て削除しました", "OK", {
-      duration: 3000,
-      verticalPosition: 'top', // 'top' | 'bottom'
-      horizontalPosition: 'center', //'start' | 'center' | 'end' | 'left' | 'right'
-      panelClass: 'pick-snackbar',
+      // ダイアログで削除する押下時、以下の処理を実行する
+      this.storageService.deleteAllData();
+
+      // pickした予定データを全て空にする
+      this.yotei = [];
+
+      // スナックバー表示
+      this.snackBar.open("全て削除しました", "OK", {
+        duration: 3000,
+        verticalPosition: 'top', // 'top' | 'bottom'
+        horizontalPosition: 'center', //'start' | 'center' | 'end' | 'left' | 'right'
+        panelClass: 'pick-snackbar',
+      });
     });
   }
 
@@ -58,5 +76,21 @@ export class PickUpComponent implements OnInit {
       horizontalPosition: 'center', //'start' | 'center' | 'end' | 'left' | 'right'
       panelClass: 'pick-snackbar',
     });
+  }
+
+  public hoge() {
+    const dialogRef = this.dialog.open(AlertDialogComponent, {
+      width: '400px',
+      height: '250px',
+      data: {
+        title: 'Pickデータの削除',
+        message: 'Pickしたデータを全て削除してもよろしいでしょうか？'
+      },
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      console.log(`Dialog result: ${result}`);
+    });
+
   }
 }
